Add Hotels and Accommodations category (#4096)
* Add Hospitality section & IHG

This is to resurect #3902  & continue on with #4024. This updates the section & adds in the _data\hospitality.yml & _img\hospitality.yml files.  This also includes the first hospitality website  IHG.com that does not have 2FA.

* Remove hospitality.yml tabs

Replace the use of tabs with spaces.

* Remove old Yaml

Remove tfa: No, as that is no longer supported or required.

* Update ihg.png

Transparent background and shrunk image size

* Corrected Icon to match bed

Removed "hotel" icon as it is nonexistent, and replaced icon with "bed".

* Change category id & title to "Lodging"

Changed Hospitality to Lodging for both [id, title].

* Add airbnb & Hilton Honors

Add Airbnb & Hilton Honors from Other.yml to Lodging.yml

* Rename to lodging.yml

* Delete Hospitality folder & create Lodging folder with images moved

* Revert "Change category id & title to "Lodging""

This reverts commit 9306844c95eb14dbff80c411ba92d330c579d372.

* Revert "Rename to lodging.yml"

This reverts commit ad7accc8456adf4a21c4edda735c0ced964e9bdb.

* Revert "Delete Hospitality folder & create Lodging folder with images moved"

This reverts commit 562343f75204c28b1ba988a231e9a435c1f95c55.

* Revert "Add airbnb & Hilton Honors"

This reverts commit 2577989d322868fd62f23629b41895db82b4826c.

* Change to lodging, move Airbnb & hilton

- Change Hospitality to Lodging, move Airbnb & Hitlon Honors with associated images into new Lodging directories.  - Remove hospitality section & folders

* Correct tab spacing

* Correct array orientation of airbnb

Move Airbnb img to end of array.

* Rename lodging.yml to hotels.yml

* Rename section loding to hotel with new title

Update title: Hotels and Accommodations

* Create hotels

* Move hotel images

* Add Marriott & fix IHG logo